// frontend/types/Story.ts
export interface Story {

  //----------------------------------------
  // Customer Profile
  //----------------------------------------

  audience: string;

  relationship: string;

  occasion: string;

  emotion: string;

  //----------------------------------------
  // Marketing Psychology
  //----------------------------------------

  painPoint: string;

  desire: string;

  fear: string;

  motivation: string;

  //----------------------------------------
  // Messaging
  //----------------------------------------

  story: string;

  marketingAngle: string;

  headline: string;

  callToAction: string;

}